split up files and refactored directory structure
