+++ src/main/scala/de/kappmeier/asnarc/elements/ActionElement.scala
@@ -1,0 +1,8 @@
+package de.kappmeier.asnarc.elements
+
+import de.kappmeier.asnarc.entity.Entity
+
+/**
+  * Elements that are updated in each iteration.
+  */
+trait ActionElement extends Entity with Element
