+++ src/main/scala/de/kappmeier/asnarc/elements/Food.scala
@@ -1,0 +1,16 @@
+package de.kappmeier.asnarc.elements
+
+import de.kappmeier.asnarc.board.Direction.Direction
+import de.kappmeier.asnarc.board.Point
+import de.kappmeier.asnarc.entity.Entity
+import de.kappmeier.asnarc.game.SnakeGame
+import de.kappmeier.asnarc.transitions.{AppendSnake, EatFood, MoveHead, StateTransition}
+
+import scala.collection.immutable.Set
+
+case class Food(p: Point) extends Entity with StaticElement {
+  override def connects = Set.empty[Direction]
+
+  def update(game: SnakeGame): Seq[StateTransition] =
+    Seq(MoveHead(game.direction()), EatFood(this), AppendSnake(game.state.player))
+}
