+++ src/main/scala/de/kappmeier/asnarc/entity/Entity.scala
@@ -1,0 +1,3 @@
+package de.kappmeier.asnarc.entity
+
+trait Entity
