+++ src/main/scala/de/kappmeier/asnarc/entity/TimedEntity.scala
@@ -1,0 +1,10 @@
+package de.kappmeier.asnarc.entity
+
+import de.kappmeier.asnarc.game.SnakeGame
+import de.kappmeier.asnarc.transitions.StateTransition
+
+trait TimedEntity {
+  val time: Int
+
+  def update(game: SnakeGame): Seq[StateTransition]
+}
