+++ src/main/scala/de/kappmeier/asnarc/event/Nothing.scala
@@ -1,0 +1,6 @@
+package de.kappmeier.asnarc.event
+
+/**
+  * A simple event that indicates no action.
+  */
+object Nothing extends SnakeEvent
