+++ src/main/scala/de/kappmeier/asnarc/event/SnakeEvent.scala
@@ -1,0 +1,3 @@
+package de.kappmeier.asnarc.event
+
+trait SnakeEvent
