+++ src/main/scala/de/kappmeier/asnarc/event/ToEvent.scala
@@ -1,0 +1,7 @@
+package de.kappmeier.asnarc.event
+
+import de.kappmeier.asnarc.entity.Entity
+
+trait ToEvent extends SnakeEvent {
+  def applyTo(entity: Entity): (Entity, Seq[SnakeEvent])
+}
