+++ src/main/scala/de/kappmeier/asnarc/game/AsnarcState.scala
@@ -1,0 +1,5 @@
+package de.kappmeier.asnarc.game
+
+import de.kappmeier.asnarc.elements.Player
+
+case class SnakeGameState(player: Player, dead: Boolean)
