+++ src/main/scala/de/kappmeier/asnarc/transitions/Death.scala
@@ -1,0 +1,9 @@
+package de.kappmeier.asnarc.transitions
+
+import de.kappmeier.asnarc.game.SnakeGame
+
+case class Death() extends StateTransition with WorldTransition {
+  def updateWorld(game: SnakeGame) {
+    game.state = game.state.copy(dead = true)
+  }
+}
