+++ src/main/scala/de/kappmeier/asnarc/transitions/MoveSnake.scala
@@ -1,0 +1,11 @@
+package de.kappmeier.asnarc.transitions
+
+import de.kappmeier.asnarc.game.SnakeGame
+
+case class MoveSnake() extends StateTransition with WorldTransition {
+  def updateWorld(game: SnakeGame) {
+    if (game.snake.nonEmpty) {
+      game.removeElement(game.snake.dequeue().p)
+    }
+  }
+}
