+++ src/main/scala/de/kappmeier/asnarc/transitions/StateTransition.scala
@@ -1,0 +1,3 @@
+package de.kappmeier.asnarc.transitions
+
+trait StateTransition
