+++ src/main/scala/de/kappmeier/asnarc/transitions/WorldTransition.scala
@@ -1,0 +1,11 @@
+package de.kappmeier.asnarc.transitions
+
+import de.kappmeier.asnarc.entity.Entity
+import de.kappmeier.asnarc.game.SnakeGame
+
+/**
+  * Events that spawn new [[Entity]]s
+  */
+trait WorldTransition extends StateTransition {
+  def updateWorld(game: SnakeGame)
+}
